md5.c now no longer uses old-style C code for declaring parameters to functions...
git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@1214 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/engine/smap.cpp
@@ -193,8 +193,8 @@ SMAP::RenderLayer(int i, bool solid, int camera_x, int camera_y, int& offset_x, 
 
   const int tile_width = m_Map.GetTileset().GetTileWidth();
   const int tile_height = m_Map.GetTileset().GetTileHeight();
-  int blit_width  = tile_width  * m_LayerInfo[i].zoomFactorX;
-  int blit_height = tile_height * m_LayerInfo[i].zoomFactorY;
+  int blit_width  = (int)(tile_width  * m_LayerInfo[i].zoomFactorX);
+  int blit_height = (int)(tile_height * m_LayerInfo[i].zoomFactorY);
 
   //m_LayerInfo[i].angle = 3.14;
